+++ test/FlashMessagesTest.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-flash for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-flash/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace ZendTest\Expressive\Flash;
 
@@ -51,6 +53,7 @@ class FlashMessagesTest extends TestCase
 
         $this->assertSame('value1', $flash->getFlash('test'));
         $this->assertSame('value2', $flash->getFlash('test-2'));
+        $this->assertSame(['test' => 'value1', 'test-2' => 'value2'], $flash->getFlashes());
     }
 
     public function testCreationAggregatesPersistsItemsWithMultipleHopsInSessionWithDecrementedHops()
@@ -83,6 +86,7 @@ class FlashMessagesTest extends TestCase
 
         $this->assertSame('value1', $flash->getFlash('test'));
         $this->assertSame('value2', $flash->getFlash('test-2'));
+        $this->assertSame(['test' => 'value1', 'test-2' => 'value2'], $flash->getFlashes());
     }
 
     public function testFlashingAValueMakesItAvailableInNextSessionButNotFlashMessages()
@@ -106,6 +110,7 @@ class FlashMessagesTest extends TestCase
         $flash->flash('test', 'value');
 
         $this->assertNull($flash->getFlash('test'));
+        $this->assertSame([], $flash->getFlashes());
     }
 
     public function testFlashNowMakesValueAvailableBothInNextSessionAndCurrentFlashMessages()
@@ -129,6 +134,7 @@ class FlashMessagesTest extends TestCase
         $flash->flashNow('test', 'value');
 
         $this->assertSame('value', $flash->getFlash('test'));
+        $this->assertSame(['test' => 'value'], $flash->getFlashes());
     }
 
     public function testProlongFlashAddsCurrentMessagesToNextSession()
@@ -178,6 +184,7 @@ class FlashMessagesTest extends TestCase
 
         $this->assertSame('value1', $flash->getFlash('test'));
         $this->assertSame('value2', $flash->getFlash('test-2'));
+        $this->assertSame(['test' => 'value1', 'test-2' => 'value2'], $flash->getFlashes());
 
         $flash->prolongFlash();
     }
@@ -217,6 +224,7 @@ class FlashMessagesTest extends TestCase
 
         $this->assertSame('value1', $flash->getFlash('test'));
         $this->assertSame('value2', $flash->getFlash('test-2'));
+        $this->assertSame(['test' => 'value1', 'test-2' => 'value2'], $flash->getFlashes());
 
         $flash->prolongFlash();
     }
@@ -252,6 +260,7 @@ class FlashMessagesTest extends TestCase
 
         $this->assertSame('value1', $flash->getFlash('test'));
         $this->assertSame('value2', $flash->getFlash('test-2'));
+        $this->assertSame(['test' => 'value1', 'test-2' => 'value2'], $flash->getFlashes());
         $flash->clearFlash();
     }
 }