+++ test/TestAsset/FlashMessages.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-flash for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-flash/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace ZendTest\Expressive\Flash\TestAsset;
 
@@ -38,6 +40,11 @@ class FlashMessages implements FlashMessagesInterface
 
     public function getFlash(string $key, $default = null)
     {
+    }
+
+    public function getFlashes() : array
+    {
+        return [];
     }
 
     public function clearFlash() : void