+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.0.0 - 2018-03-15
+
+### Added
+
+- [#5](https://github.com/zendframework/zend-expressive-flash/pull/5) adds
+  support for PSR-15 interfaces.
+
+- [#6](https://github.com/zendframework/zend-expressive-flash/pull/6) adds
+  the FlashMessages::getFlashes() method to retrievs an array with all
+  messages for the current request.
+
+### Changed
+
+- Nothing.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- [#3](https://github.com/zendframework/zend-expressive-flash/pull/3) and
+  [#5](https://github.com/zendframework/zend-expressive-flash/pull/5) remove
+  support for http-interop/http-middleware and
+  http-interop/http-server-middleware.
+
+### Fixed
+
+- Nothing.
+
 ## 0.1.0 - 2017-10-10
 
 Initial release.